[api] Move terminator instructions into own trait
