+++ src/main/kotlin/dev/supergrecko/vexe/llvm/ir/instructions/BrInstruction.kt
@@ -3,11 +3,11 @@ package dev.supergrecko.vexe.llvm.ir.instructions
 import dev.supergrecko.vexe.llvm.internal.util.fromLLVMBool
 import dev.supergrecko.vexe.llvm.ir.Instruction
 import dev.supergrecko.vexe.llvm.ir.Value
+import dev.supergrecko.vexe.llvm.ir.instructions.traits.Terminator
 import org.bytedeco.llvm.LLVM.LLVMValueRef
 import org.bytedeco.llvm.global.LLVM
 
-public class BrInstruction internal constructor() :
-    Instruction() {
+public class BrInstruction internal constructor() : Instruction(), Terminator {
     public constructor(llvmRef: LLVMValueRef) : this() {
         ref = llvmRef
     }
