+++ src/main/kotlin/dev/supergrecko/vexe/llvm/ir/instructions/CatchSwitchInstruction.kt
@@ -3,12 +3,14 @@ package dev.supergrecko.vexe.llvm.ir.instructions
 import dev.supergrecko.vexe.llvm.internal.util.map
 import dev.supergrecko.vexe.llvm.ir.BasicBlock
 import dev.supergrecko.vexe.llvm.ir.Instruction
+import dev.supergrecko.vexe.llvm.ir.instructions.traits.Terminator
 import org.bytedeco.javacpp.PointerPointer
 import org.bytedeco.llvm.LLVM.LLVMBasicBlockRef
 import org.bytedeco.llvm.LLVM.LLVMValueRef
 import org.bytedeco.llvm.global.LLVM
 
-public class CatchSwitchInstruction internal constructor() : Instruction() {
+public class CatchSwitchInstruction internal constructor() : Instruction(),
+    Terminator {
     public constructor(llvmRef: LLVMValueRef) : this() {
         ref = llvmRef
     }
