+++ src/main/kotlin/dev/supergrecko/vexe/llvm/ir/instructions/CleanupRetInstruction.kt
@@ -1,9 +1,11 @@
 package dev.supergrecko.vexe.llvm.ir.instructions
 
 import dev.supergrecko.vexe.llvm.ir.Instruction
+import dev.supergrecko.vexe.llvm.ir.instructions.traits.Terminator
 import org.bytedeco.llvm.LLVM.LLVMValueRef
 
-public class CleanupRetInstruction internal constructor() : Instruction() {
+public class CleanupRetInstruction internal constructor() : Instruction(),
+    Terminator {
     public constructor(llvmRef: LLVMValueRef) : this() {
         ref = llvmRef
     }
