+++ src/main/kotlin/dev/supergrecko/vexe/llvm/ir/instructions/InvokeInstruction.kt
@@ -4,11 +4,12 @@ import dev.supergrecko.vexe.llvm.ir.BasicBlock
 import dev.supergrecko.vexe.llvm.ir.Instruction
 import dev.supergrecko.vexe.llvm.ir.Value
 import dev.supergrecko.vexe.llvm.ir.instructions.traits.CallBase
+import dev.supergrecko.vexe.llvm.ir.instructions.traits.Terminator
 import org.bytedeco.llvm.LLVM.LLVMValueRef
 import org.bytedeco.llvm.global.LLVM
 
 public class InvokeInstruction internal constructor() : Instruction(),
-    CallBase {
+    CallBase, Terminator {
     public constructor(llvmRef: LLVMValueRef) : this() {
         ref = llvmRef
     }
